Use `pytest_pyfunc_call` hook to intercept figure
Instead modifying the test function itself by wrapping it in a function which runs the tests, use pytest hooks to intercept the generated figure and then run the tests. This should be a more robust approach that doesn't need as many special cases to be hardcoded.

I have also refactored the get_marker and get_compare functions/methods to simplify these.