Raise error if multiple tests generate baseline at same path
By default, baseline images are just given the same name as the test. However, there can be duplicates if the same test name exists in multiple modules or if the same test name is within different testing classes.

This keeps a record of each path written to while generating baseline images, and before saving the image it double checks that an image hasn't already been written to there in the current pytest run.